refactor shielding to a hook
