+++ components/TxForm.tsx
@@ -1,51 +1,27 @@
 import { useToken } from "@/contexts/TokenContext";
 import useNotifications from "@/hooks/useNotifications";
-import useShieldPrivateKey from "@/hooks/useShieldPrivateKey";
+import useRailgunTx from "@/hooks/useRailgunTx";
 import { Button } from "@chakra-ui/button";
 import { FormControl, FormLabel } from "@chakra-ui/form-control";
 import { Input, InputGroup, InputRightElement } from "@chakra-ui/input";
 import { Box, Flex } from "@chakra-ui/layout";
 import { Select } from "@chakra-ui/select";
-import {
-  gasEstimateForShield,
-  gasEstimateForShieldBaseToken,
-  getRailgunSmartWalletContractForNetwork,
-  populateShield,
-  populateShieldBaseToken,
-} from "@railgun-community/quickstart";
-import {
-  RailgunERC20AmountRecipient,
-  RailgunERC20Amount,
-  NetworkName,
-  TransactionGasDetailsSerialized,
-  EVMGasType,
-  NETWORK_CONFIG,
-  deserializeTransaction,
-} from "@railgun-community/shared-models";
+import { getRailgunSmartWalletContractForNetwork } from "@railgun-community/quickstart";
+import { NetworkName } from "@railgun-community/shared-models";
 import { erc20ABI } from "@wagmi/core";
 import { ethers, constants, BigNumber } from "ethers";
 import { useState } from "react";
-import {
-  useAccount,
-  useSigner,
-  usePrepareContractWrite,
-  useContractWrite,
-  useProvider,
-} from "wagmi";
+import { usePrepareContractWrite, useContractWrite } from "wagmi";
 
 export const TxForm = () => {
   // TODO: Placeholder notification for shielding
-  const { tokenList, tokenAllowances, weth } = useToken();
+  const { tokenList, tokenAllowances } = useToken();
   const network = NetworkName.EthereumGoerli;
   const { txNotify, notifyUser } = useNotifications();
   let abi = [
     "function transfer(address,uint256) returns (bool)",
     "function approve(address,uint256) returns (bool)",
   ];
-  const { isConnected, address } = useAccount();
-  const { data: signer } = useSigner();
-  const provider = useProvider();
-  const { getShieldPrivateKey } = useShieldPrivateKey();
   const [tokenAddress, setTokenAddress] = useState<string>();
   const [tokenAmount, setTokenAmount] = useState<string>("");
   const [tokenDecimals, setTokenDecimals] = useState<number>();
@@ -62,6 +38,7 @@ export const TxForm = () => {
   const [recipient, setRecipient] = useState<string>(
     "0zk1qyn0qa5rgk7z2l8wyncpynmydgj7ucrrcczhl8k27q2rw5ldvv2qrrv7j6fe3z53ll5j4fjs9j5cmq7mxsaulah7ykk6jwqna3nwvxudp5w6fwyg8cgwkwwv3g4"
   );
+  const { shield } = useRailgunTx();
 
   const needsApproval = ethers.utils
     .parseUnits(tokenAmount || "0", tokenDecimals)
@@ -69,135 +46,16 @@ export const TxForm = () => {
 
   const doSubmit: React.FormEventHandler = async (e) => {
     // TODO: Form validation
-
-    // Public wallet to shield from.
-    const fromWalletAddress = address as `0x{string}`;
-
-    // The shieldPrivateKey enables the sender to decrypt
-    // the receiver's address in the future.
-    const shieldPrivateKey = await getShieldPrivateKey();
-
-    // baseToken
-    if (tokenAddress === ethers.utils.getAddress("0x" + "e".repeat(40))) {
-      const wrappedERC20Amount: RailgunERC20Amount = {
-        tokenAddress: weth!, // wETH
-        amountString: ethers.utils
-          .parseUnits(tokenAmount!, tokenDecimals)
-          .toHexString(), // hexadecimal amount
-      };
-
-      const { gasEstimateString, error: err } =
-        await gasEstimateForShieldBaseToken(
-          network,
-          recipient,
-          shieldPrivateKey,
-          wrappedERC20Amount,
-          fromWalletAddress
-        );
-      if (err) {
-        // Handle gas estimate error.
-      }
-
-      const { maxFeePerGas, maxPriorityFeePerGas } =
-        await provider.getFeeData();
-
-      const gasDetailsSerialized: TransactionGasDetailsSerialized = {
-        evmGasType: EVMGasType.Type2, // Depends on the chain (BNB uses type 0)
-        gasEstimateString: gasEstimateString!, // Output from gasEstimateForShieldBaseToken
-        maxFeePerGasString: maxFeePerGas!.toHexString(), // Current gas Max Fee
-        maxPriorityFeePerGasString: maxPriorityFeePerGas!.toHexString(), // Current gas Max Priority Fee
-      };
-
-      const { serializedTransaction, error } = await populateShieldBaseToken(
-        network,
-        recipient,
-        shieldPrivateKey,
-        wrappedERC20Amount,
-        gasDetailsSerialized
-      );
-      if (error) {
-        throw error;
-      }
-
-      const { chain } = NETWORK_CONFIG[network];
-
-      const transactionRequest: ethers.providers.TransactionRequest =
-        deserializeTransaction(
-          serializedTransaction!,
-          undefined, // nonce (optional)
-          chain.id
-        );
-
-      // Public wallet to shield from.
-      transactionRequest.from = address;
-
-      const tx = await signer?.sendTransaction(transactionRequest);
-      await tx?.wait().then(() => {
-        txNotify(tx.hash);
-      });
-      return;
-    }
-
-    // else: ERC20
-
-    // Formatted token amounts and recipients.
-    const erc20AmountRecipients: RailgunERC20AmountRecipient[] = [
-      {
-        tokenAddress: tokenAddress!,
-        amountString: ethers.utils
-          .parseUnits(tokenAmount!, tokenDecimals)
-          .toHexString(), // must be hex
-        recipientAddress: recipient!, // RAILGUN address
-      },
-    ];
-
-    const { gasEstimateString, error: err } = await gasEstimateForShield(
-      network,
-      shieldPrivateKey,
-      erc20AmountRecipients,
-      [], // nftAmountRecipients
-      fromWalletAddress
-    );
-
-    if (err) {
-      throw err;
-    }
-
-    const { maxFeePerGas, maxPriorityFeePerGas } = await provider.getFeeData();
-    const gasDetailsSerialized: TransactionGasDetailsSerialized = {
-      evmGasType: EVMGasType.Type2, // Depends on the chain (BNB uses type 0)
-      gasEstimateString: gasEstimateString!, // Output from gasEstimateForShield
-      maxFeePerGasString: maxFeePerGas!.toHexString(), // Current gas Max Fee
-      maxPriorityFeePerGasString: maxPriorityFeePerGas!.toHexString(), // Current gas Max Priority Fee
-    };
-
-    const { serializedTransaction, error } = await populateShield(
-      network,
-      shieldPrivateKey,
-      erc20AmountRecipients,
-      [], // nftAmountRecipients
-      gasDetailsSerialized
-    );
-    if (error) {
-      throw error;
-    }
-
-    const { chain } = NETWORK_CONFIG[network];
-
-    const transactionRequest: ethers.providers.TransactionRequest =
-      deserializeTransaction(
-        serializedTransaction as string,
-        undefined, // nonce (optional)
-        chain.id
-      );
-
-    // Public wallet to shield from.
-    transactionRequest.from = address;
-    // TODO: handle transaction status and notifications in a feature
-    const tx = await signer?.sendTransaction(transactionRequest);
-    await tx?.wait().then(() => {
-      txNotify(tx.hash);
+    if (!tokenAddress || !tokenAmount || !tokenDecimals || !recipient)
+      throw new Error("bad form");
+    const tx = await shield({
+      tokenAddress,
+      tokenAmount,
+      tokenDecimals,
+      recipient,
     });
+    await tx?.wait();
+    txNotify(tx!.hash);
   };
 
   return (
